requirements.txt
streamlit
tensorflow
opencv-python
Pillow
numpy
streamlit-drawable-canvas